+++ R/tar_progress_branches.R
@@ -1,7 +1,8 @@
-#' @title Read the target progress of the latest run of the pipeline.
+#' @title Tabulate the progress of dynamic branches.
 #' @export
 #' @description Read a project's target progress data for the most recent
-#'   run of [tar_make()] or similar. Only the most recent record is shown.
+#'   run of the pipeline and display the tabulated status
+#'   of dynamic branches. Only the most recent record is shown.
 #' @return A data frame with one row per target per progress status
 #'   and the following columns.
 #'   * `name`: name of the pattern.
@@ -32,7 +33,7 @@
 #' }
 tar_progress_branches <- function(names = NULL, fields = NULL) {
   assert_store()
-  assert_path(file.path("_targets/meta/progress"))
+  assert_path(path_progress())
   out <- tibble::as_tibble(progress_init()$database$read_condensed_data())
   out <- tar_progress_branches_summary(out)
   names_quosure <- rlang::enquo(names)
@@ -56,7 +57,7 @@ tar_progress_branches_summary <- function(progress) {
     progress = gsub(".* ", "", group),
     branches = as.integer(table)
   )
-  levels <- c("started", "built", "canceled", "errored")
+  levels <- c("started", "built", "errored", "canceled")
   bins <- map(levels, ~tar_progress_branches_bin(.x, long))
   out <- progress[progress$type == "pattern",, drop = FALSE] # nolint
   out <- tibble::tibble(name = out$name, branches = out$branches)
@@ -80,48 +81,5 @@ tar_progress_branches_bin <- function(level, long) {
 # Just for the tar_watch() app. # nolint
 tar_progress_branches_gt <- function() {
   progress <- tar_progress_branches(names = NULL, fields = NULL)
-  out <- gt_borderless(progress)
-  out <- gt::tab_style(
-    out,
-    gt::cell_text(weight = "bold"),
-    locations = gt::cells_column_labels(everything())
-  )
-  colors <- data_frame(
-    progress = c("started", "built", "canceled", "errored"),
-    fill = c("#DC863B", "#E1BD6D", "#FAD510", "#C93312"),
-    color = c("black", "black", "black", "white")
-  )
-  for (index in seq_len(nrow(colors))) {
-    out <- gt::tab_style(
-      out,
-      style = list(
-        gt::cell_fill(color = colors$fill[index]),
-        gt::cell_text(color = colors$color[index])
-      ),
-      locations = gt::cells_body(
-        columns = colors$progress[index],
-        rows = progress[[colors$progress[index]]] > 0L
-      )
-    )
-  }
-  out
-}
-
-gt_borderless <- function(x) {
-  out <- gt::gt(x)
-  out <- gt::cols_width(out, everything() ~ gt::pct(16.6))
-  out <- gt::cols_align(out, align = "left", columns = everything())
-  out <- gt::tab_options(
-    out,
-    row.striping.include_table_body = TRUE,
-    table.border.top.style = "hidden",
-    table.border.bottom.style = "hidden",
-    table_body.border.bottom.style = "hidden"
-  )
-  out <- gt::tab_style(
-    out,
-    style = gt::cell_borders(weight = gt::px(0)),
-    locations = gt::cells_body()
-  )
-  out
+  tar_progress_display_gt(progress)
 }