+++ NEWS.md
@@ -3,6 +3,13 @@
 ## New features
 
 * Implement `tar_timestamp()` to get the last modified timestamp of a target's data (#378).
+* Implement `tar_progress_summary()` to compactly summarize all pipeline progress (#380).
+* Add a `characters` argument of `tar_traceback()` to cap the traceback line lengths (#383).
+* Add new "summary" and "about" views to `tar_watch()` (#382).
+
+## Enhancements
+
+* Attempt to print only the useful part of the traceback in `tar_traceback()` (#383).
 
 # targets 0.3.1